Backend/requirements.txt
python-dotenv>=1.0.0
pinecone>=5.0.0
sentence-transformers>=2.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
